+++ _book/docs/exemplos/minhaprimeiraref.bib
@@ -111,4 +111,20 @@
   url              = {https://CRAN.R-project.org/package=dplyr},
 }
 
+@Article{ipyrad,
+  author           = {Eaton, Deren A R and Overcast, Isaac},
+  journal          = {Bioinformatics},
+  title            = {{ipyrad: Interactive assembly and analysis of RADseq datasets}},
+  year             = {2020},
+  issn             = {1367-4803},
+  month            = {01},
+  note             = {btz966},
+  abstract         = {{ipyrad is a free and open source tool for assembling and analyzing restriction site-associated DNA sequence datasets using de novo and/or reference-based approaches. It is designed to be massively scalable to hundreds of taxa and thousands of samples, and can be efficiently parallelized on high performance computing clusters. It is available both as a command line interface and as a Python package with an application programming interface, the latter of which can be used interactively to write complex, reproducible scripts and implement a suite of downstream analysis tools.ipyrad is a free and open source program written in Python. Source code is available from the GitHub repository (https://github.com/dereneaton/ipyrad/), and Linux and MacOS installs are distributed through the conda package manager. Complete documentation, including numerous tutorials, and Jupyter notebooks demonstrating example assemblies and applications of downstream analysis tools are available online: https://ipyrad.readthedocs.io/.}},
+  doi              = {10.1093/bioinformatics/btz966},
+  eprint           = {https://academic.oup.com/bioinformatics/advance-article-pdf/doi/10.1093/bioinformatics/btz966/31875323/btz966.pdf},
+  file             = {:C\:/Users/ricop/OneDrive/Documentos/DOC/PROJETO\_DOC/PAPERS_SUPORTE_TESE/EatonOvercast2020.pdf:PDF},
+  modificationdate = {2024-07-27T11:02:29},
+  url              = {https://doi.org/10.1093/bioinformatics/btz966},
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
