+++ _book/docs/exemplos/referencia_cruzada.Rmd
@@ -3,26 +3,77 @@ title: "Análise exploratória de dados"
 subtitle: "Aqui vai o subtítulo"
 author: Ricardinho
 date: 4 Jun 2024
+# documentclass: book
 output: 
   bookdown::pdf_document2:
     toc: true
-    toc-title: Sumário
+    # includes:
+    #   in_header: "header.tex"
+    #   before_body: "before-body.tex"
+  # bookdown::word_document2:
+  #   reference_docx: modelo.docx
+  #   theme: cerulean
+  #   fig_width: 12
+  #   fig_height: 10
+# lof: yes
+# lot: yes
 bibliography: minhaprimeiraref.bib
+csl: associacao-brasileira-de-normas-tecnicas.csl
 ---
 
-```{r}
+```{r, echo=FALSE, eval=TRUE, message=FALSE}
 library("knitr")
+variavel1 <- 5+2
+variavel2 <- variavel1 +2
+
+library("dplyr")
+
+exemplinho <- read.table("exemplinho.csv", sep = ",", header = TRUE)
 ```
 
-## Introdução
+\newpage
+
+```{r, echo=FALSE, eval=FALSE}
+1+1+1+1+1+2
+```
+
+
+```{r}
+#| echo: true
+#| eval: false
+#| include: false
+#| message: false
+#| warning: false
+5+2
+```
+
+# Introdução
+
+## Subintrodução
+
+Aqui está a variável 1: `r variavel1`.  
+
 
 Aqui eu tenho um texto introdutório sobre minha análise exploratória de dados (AED) referentes a carros.  
 
 Vejam a figura \@ref(fig:fig1).  
 
-```{r fig1, fig.cap = "Aqui vai a legenda das figuras."}
+```{r fig1, fig.cap = "Aqui vai a legenda das figuras.", out.width="120%"}
 plot(mtcars)
 ```
+
+Olha a fotinha \@ref(fig:fotinha).  
+
+```{r fotinha, fig.cap= "Fotinha", out.width="50%"}
+#| echo: false
+knitr::include_graphics("fotinha.JPG")
+```
+
+```{r}
+magick::image_read("../fotinha.JPG") |> 
+  magick::image_scale("x700")
+```
+
 
 A tabela \@ref(tab:minha-tabela) expõe as primeiras 6 linhas do conjunto de dados 'mtcars'.  
 
@@ -30,11 +81,70 @@ A tabela \@ref(tab:minha-tabela) expõe as primeiras 6 linhas do conjunto de dad
 knitr::kable(head(mtcars), caption = "Aqui vai a legenda de tabelas, isto é, dentro do argumento 'caption' da função knitr::kable")
 ```
 
-## Material e métodos
+
+
+# Material e métodos
 
 Aqui eu discorro sobre os métodos utilizados para responder minhas perguntas.  
 
 Uma lista de refs [@Xieetal2020;@Yaoetal2024].  
 
-## Referências
+
+## Localidades
+
+## Análise de dados
+
+```{r, echo=FALSE, warning=FALSE}
+library("palmerpenguins")
+library("ggplot2")
+ggplot(penguins) +
+  geom_point(aes(x = bill_length_mm, y = bill_depth_mm, color = species)) +
+  theme_bw()
+
+```
+
+# Resultados
+
+## Genética
+
+### ddradSEQ
+
+## Morfologia
+
+Reunimos um conjunto de `r nrow(mtcars)` observações e `r ncol(mtcars)` variáveis.  
+
+```{r}
+mtcars2 <- dplyr::bind_rows(mtcars, mtcars)
+```
+
+Na série temporal do ano de 2019, houve `r nrow(mtcars2)` observações e `r ncol(mtcars2)` variáveis para estimar o traço funcional do solo papapa.  
+
+A tabela \@ref(tab:tab2) evidencia as seis primeiras linhas de nosso conjunto de dados.  
+
+```{r tab2}
+#| echo: false
+dados = read.table("exemplinho.csv", header = TRUE, sep = ",")
+knitr::kable(dados, caption = "Seis primeiras linhas do conjunto de dados iris.")
+```
+
+Coletamos `r length(unique(dados$Species))` espécies diferentes.  
+
+As espécies coletadas foram `r glue::glue_collapse(x = sort(unique(dados$Species)), sep = ", ", last = " e ")`.  
+
+`r 1+1`
+
+| Nome    | Altura |
+|---------|--------|
+| Lysne   | `r 1+1`|
+| Luciana | `r 1+2`|
+| Rhaylla | 1,68   |
+| José    | 1,82   |
+| Sylvio  | 1,73   |
+| Carlos  | 1,75   |
+| Ricardo | 1,79   |
+
+: https://www.tablesgenerator.com/
+
+
+# Referências
