+++ _book/site_libs/quarto-html/quarto-syntax-highlighting-9707efa71df770813d83ac86ec1d4592.css
@@ -36,94 +36,8 @@
   --quarto-font-monospace: SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace;
 }
 
-code span.al {
-  font-weight: bold;
-  color: #ff5555;
-}
-
-code span.an {
-  color: #6a737d;
-}
-
-code span.at {
-  color: #d73a49;
-}
-
-code span.bn {
-  color: #005cc5;
-}
-
-code span.bu {
-  color: #d73a49;
-}
-
-code span.ch {
-  color: #032f62;
-}
-
-code span.co {
-  color: #6a737d;
-}
-
-code span.cv {
-  color: #6a737d;
-}
-
-code span.cn {
-  color: #005cc5;
-}
-
-code span.cf {
-  color: #d73a49;
-}
-
-code span.dt {
-  color: #d73a49;
-}
-
-code span.dv {
-  color: #005cc5;
-}
-
-code span.do {
-  color: #6a737d;
-}
-
-code span.er {
-  color: #ff5555;
-  text-decoration: underline;
-}
-
-code span.ex {
-  font-weight: bold;
-  color: #d73a49;
-}
-
-code span.fl {
-  color: #005cc5;
-}
-
-code span.fu {
-  color: #6f42c1;
-}
-
-code span.im {
-  color: #032f62;
-}
-
-code span.in {
-  color: #6a737d;
-}
-
-code span.kw {
-  color: #d73a49;
-}
-
+/* syntax highlight based on Pandoc's rules */
 pre > code.sourceCode > span {
-  color: #24292e;
-}
-
-code span {
   color: #24292e;
 }
 
@@ -136,38 +50,155 @@ div.sourceCode pre.sourceCode {
   color: #24292e;
 }
 
+/* Normal */
+code span {
+  color: #24292e;
+}
+
+/* Alert */
+code span.al {
+  font-weight: bold;
+  color: #ff5555;
+}
+
+/* Annotation */
+code span.an {
+  color: #6a737d;
+}
+
+/* Attribute */
+code span.at {
+  color: #d73a49;
+}
+
+/* BaseN */
+code span.bn {
+  color: #005cc5;
+}
+
+/* BuiltIn */
+code span.bu {
+  color: #d73a49;
+}
+
+/* ControlFlow */
+code span.cf {
+  color: #d73a49;
+}
+
+/* Char */
+code span.ch {
+  color: #032f62;
+}
+
+/* Constant */
+code span.cn {
+  color: #005cc5;
+}
+
+/* Comment */
+code span.co {
+  color: #6a737d;
+}
+
+/* CommentVar */
+code span.cv {
+  color: #6a737d;
+}
+
+/* Documentation */
+code span.do {
+  color: #6a737d;
+}
+
+/* DataType */
+code span.dt {
+  color: #d73a49;
+}
+
+/* DecVal */
+code span.dv {
+  color: #005cc5;
+}
+
+/* Error */
+code span.er {
+  color: #ff5555;
+  text-decoration: underline;
+}
+
+/* Extension */
+code span.ex {
+  font-weight: bold;
+  color: #d73a49;
+}
+
+/* Float */
+code span.fl {
+  color: #005cc5;
+}
+
+/* Function */
+code span.fu {
+  color: #6f42c1;
+}
+
+/* Import */
+code span.im {
+  color: #032f62;
+}
+
+/* Information */
+code span.in {
+  color: #6a737d;
+}
+
+/* Keyword */
+code span.kw {
+  color: #d73a49;
+}
+
+/* Operator */
 code span.op {
   color: #24292e;
 }
 
+/* Preprocessor */
 code span.pp {
   color: #d73a49;
 }
 
+/* RegionMarker */
 code span.re {
   color: #6a737d;
 }
 
+/* SpecialChar */
 code span.sc {
   color: #005cc5;
 }
 
+/* SpecialString */
 code span.ss {
   color: #032f62;
 }
 
+/* String */
 code span.st {
   color: #032f62;
 }
 
+/* Variable */
 code span.va {
   color: #e36209;
 }
 
+/* VerbatimString */
 code span.vs {
   color: #032f62;
 }
 
+/* Warning */
 code span.wa {
   color: #ff5555;
 }
@@ -176,4 +207,4 @@ code span.wa {
   content: "</";
 }
 
-/*# sourceMappingURL=debc5d5d77c3f9108843748ff7464032.css.map */
+/*# sourceMappingURL=144a707635c52f7585e125478c4c7d06.css.map */
