+++ _freeze/04_compilacao/execute-results/html.json
@@ -2,7 +2,7 @@
   "hash": "da9ba698431007146ee6b0e3fd8c63e6",
   "result": {
     "engine": "knitr",
-    "markdown": "---\ntitle: \"Compilação de arquivos Rmd\"\nformat: html\naliases: \n  - /compilacao/\n---\n\n\n\n\n## Objetivos de aprendizado\n\n+ Ser capaz de renderizar um arquivo Rmd em diferentes formatos de saída;\n+ Conhecer os comandos e opções disponíveis para a renderização de documentos.\n\n## O que significa compilar um arquivo Rmd?\n\n![Fluxo de trabalho que ocorre durante a compilação de um arquivo Rmd. Extraído do sítio digital do livro \"Rmarkdown Cookbook\".](https://bookdown.org/yihui/rmarkdown-cookbook/images/workflow.png)\n\n## Compilação utilizando a estrutura do RStudio {#compilarrmd}\n\nO RStudio é otimizado para lidar com arquivos Rmd.\nPara compilar um arquivo neste programa, basta abrir o arquivo Rmd e clicar no botão \"Knit\" na barra de ferramentas (@fig-botaoknit).\nIsso compilará o arquivo Rmd, executando o código (se presente!) e incorporando os resultados no documento final.  \n\n![Botão \"Knit\" destacado com um retângulo verde. Este botão está localizado em uma barra abaixo dos nomes de arquivos abertos dentro do RStudio. Clicando neste botão, você compila o arquivo para o formato especificado no próprio arquivo. Caso clique em uma seta ao lado do nome \"Knit\", outras opções de formato final são apresentadas, como mostra esta figura.](figuras/botao_knit.png){#fig-botaoknit}\n\n### Linha de comando\n\nPara compilar um arquivo Rmd usando a linha de comando, você pode usar o seguinte comando no console do R:\n\n```r\nrmarkdown::render(\"caminho/do/seu/arquivo.Rmd\")\n```\n\nPor exemplo, vamos imaginar que tenhamos um arquivo chamado \"arq1.Rmd\" em minha pasta de trabalho e desejo compilá-lo.\nDevemos executar o seguinte comando:\n\n\n\n\n::: {.cell}\n\n```{.r .cell-code}\nrmarkdown::render(\"arq1.Rmd\")\n```\n:::\n\n\n\n\n\n::: {.callout-tip}\nCaro leitor,\n\nCaso você não tenha conhecimento suficiente em R ou se sinta inseguro ao utilizar a linguagem, recomendo a leitura do livro [Curso básico de introdução à linguagem R](https://intror.netlify.app/), disponível gratuitamente na rede.\n:::\n\n\n\n",
+    "markdown": "---\ntitle: \"Compilação de arquivos Rmd\"\nformat: html\naliases: \n  - /compilacao/\n---\n\n## Objetivos de aprendizado\n\n+ Ser capaz de renderizar um arquivo Rmd em diferentes formatos de saída;\n+ Conhecer os comandos e opções disponíveis para a renderização de documentos.\n\n## O que significa compilar um arquivo Rmd?\n\n![Fluxo de trabalho que ocorre durante a compilação de um arquivo Rmd. Extraído do sítio digital do livro \"Rmarkdown Cookbook\".](https://bookdown.org/yihui/rmarkdown-cookbook/images/workflow.png)\n\n## Compilação utilizando a estrutura do RStudio {#compilarrmd}\n\nO RStudio é otimizado para lidar com arquivos Rmd.\nPara compilar um arquivo neste programa, basta abrir o arquivo Rmd e clicar no botão \"Knit\" na barra de ferramentas (@fig-botaoknit).\nIsso compilará o arquivo Rmd, executando o código (se presente!) e incorporando os resultados no documento final.  \n\n![Botão \"Knit\" destacado com um retângulo verde. Este botão está localizado em uma barra abaixo dos nomes de arquivos abertos dentro do RStudio. Clicando neste botão, você compila o arquivo para o formato especificado no próprio arquivo. Caso clique em uma seta ao lado do nome \"Knit\", outras opções de formato final são apresentadas, como mostra esta figura.](figuras/botao_knit.png){#fig-botaoknit}\n\n### Linha de comando\n\nPara compilar um arquivo Rmd usando a linha de comando, você pode usar o seguinte comando no console do R:\n\n```r\nrmarkdown::render(\"caminho/do/seu/arquivo.Rmd\")\n```\n\nPor exemplo, vamos imaginar que tenhamos um arquivo chamado \"arq1.Rmd\" em minha pasta de trabalho e desejo compilá-lo.\nDevemos executar o seguinte comando:\n\n\n::: {.cell}\n\n```{.r .cell-code}\nrmarkdown::render(\"arq1.Rmd\")\n```\n:::\n\n\n\n::: {.callout-tip}\nCaro leitor,\n\nCaso você não tenha conhecimento suficiente em R ou se sinta inseguro ao utilizar a linguagem, recomendo a leitura do livro [Curso básico de introdução à linguagem R](https://intror.netlify.app/), disponível gratuitamente na rede.\n:::\n\n\n\n",
     "supporting": [],
     "filters": [
       "rmarkdown/pagebreak.lua"
